Move resolve_reference_name into djls-semantic
Template-reference resolution (primary-name anchoring plus the reference
kind's self-reference rule) is Django meaning, not IDE behavior, so it
moves from djls-ide to djls-semantic's references module next to the
reference aggregation. With its last external caller gone,
TemplateReferenceKind::allow_self becomes private.
