+++ crates/djls-ide/src/lib.rs
@@ -11,7 +11,6 @@ mod links;
 mod navigation;
 mod snippets;
 mod symbols;
-mod templates;
 mod warmup;
 
 pub use code_actions::code_actions;
