+++ crates/djls-ide/src/navigation.rs
@@ -2,13 +2,13 @@ use djls_project::FindTemplateResult;
 use djls_project::template_resolution;
 use djls_semantic::SemanticOffsetContext;
 use djls_semantic::references_to_template_name;
+use djls_semantic::resolve_reference_name;
 use djls_source::File;
 use djls_source::Offset;
 use tower_lsp_server::ls_types;
 
 use crate::ext::SpanExt;
 use crate::ext::Utf8PathExt;
-use crate::templates::resolve_reference_name;
 
 pub fn goto_definition(
     db: &dyn djls_semantic::Db,
